Always check for collisions before joining ring
Perform collision check regardless of whether the node is bootstrapping
or not (unless replacing). This involves running a shadow round of gossip
to learn ring detailis, which is problematic if all seeds are started
concurrently. A node may now exit its shadow round if either it can
successfully gossip with a seed or it learns that all seeds are currently
in their own shadow round. Additionally, if neither of those conditions
are met before the shadow round timeout is reached and iff the node
considers itself a seed, a warning is logged and startup allowed to
continue. This is to allow a single seed to be started when all other
seeds are legitimately down.
The collision check and shadow round can be skipped completely (for
testing etc) by setting cassandra.allow_unsafe_join=true.