index: do not re-insert values in IndexInfo
On keyspace with multiple indexes we end up doing one blocking
flush per index on startup. This is by far the most expensive operation
and can make cassandra boot in minutes instead of milliseconds.

Instead, use "IF NOT EXISTS;" to avoid changes when we don't need to.

I suspect that this could also be applied to some other tables in `system`.