LIST ROLES returns Set<String> not String for DCs
The thinking behind this is that a non-cqlsh client executing a LIST ROLES
query should reasonably be able to reconstruct the role from the resultset.
As it was, the client would need to implement specific logic to translate from
the string in the 'datacenters' column to the logical set of DC names. So I've
just inverted that to have cqlsh do the mapping from set<dc name> -> human
readable string ("ALL", "n/a" etc).