Update current index block pointer when slice precedes partition start
When reverse iterating an indexed sstable partition, if the end bound of
the query slice is < the first unfiltered in the partition the current
index block boundary is not updated. This causes the reader to incorrectly
jump to the end of the partition and start reading from there once the
initial emtpy iterator has been consumed.