Restore shorter timeout for Prepare repair messages
CASSANDRA-9292 removed the 1hr timeout on prepare messages, but this
was inadvertently undone when CASSANDRA-13397 was committed. As nothing
long running is done in the repair phase anymore, this timeout can
safely be reduced.

The test change is necessary because now the repair session is correctly
reporting that it didn't receive required responses. Previously, it
was incorrectly reporting a non-success response from a specific peer.