Revert distiction between weighted and unweighted caches.
Keeps the distinction at the presentation layer (virtual tables and
nodetool), but the having subinterfaces and implementations to
distinguish un/weighted seems too much. For an unweighted cache,
weightedSize() == size(). Aside from that, only the units of
size and capacity are different and we don't care about that except in
the presentation, so it can be documented in comments.