[CEP-21] Upgrade support
Following an upgrade, nodes in an existing cluster will enter a minimal
modification mode. In this state, the set of allowed cluster metadata
modifications is constrained to include only the addition, removal and
replacement of nodes, to allow failed hosts to be replaced during the
upgrade.

In this mode the CMS has no members and each peer maintains its
own ClusterMetadata independently. This metadata is intitialised at
startup from system tables and gossip is used to propagate the permitted
metadata changes.

When the operator is ready, one node is chosen for promotion to the initial
CMS, which is done manually via nodetool. At this point, the candidate node
will propose itself as the initial CMS and attempt to gain consensus from
the rest of the cluster. If successful, it verifies that all peers have an
identical view of cluster metadata and initialises the distributed log with
a snapshot of that metadata.

Once this process is complete all future cluster metadata updates are performed
via the CMS using the global log and reverting to the previous method of
metadata management is not supported. Further members can and should be added
to the CMS via the nodetool command.

Co-authored-by: Marcus Eriksson <marcuse@apache.org>
Co-authored-by: Alex Petrov <oleksandr.petrov@gmail.com>
Co-authored-by: Sam Tunnicliffe <samt@apache.org>