[CEP-21] Adjust to new Register checks
Test setup needs to handle the improved validation when executing Register
that was introduced to support nodes changing IP (in
b940e18d8983e12bda72ca0f27982616193f6f25).

In test code, this is part of an ongoing effort to rationalise the
various test setup code in the project and there is still more to do.